[#44] refactor: change data type

+++ backend/src/main/java/com/ssafy/picple/domain/background/dto/response/openai/AIBackgroundResponse.java
@@ -10,6 +10,6 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class AIBackgroundResponse {
-	int created;
+	long created;
 	List<ImageObject> data;
 }
